+++ gradle.properties
@@ -8,7 +8,7 @@ org.gradle.jvmargs=-Xmx1G
 	loader_version=0.16.10
 
 # Mod Properties
-	mod_version=3.1.0
+	mod_version=3.2.1
 	maven_group=net.elytraautopilot
 	archives_base_name=fabric-elytra-autopilot-1.21.4
